+++ packages/code-du-travail-frontend/src/e2e/helpers/index.ts
@@ -2,3 +2,4 @@ export * from "./seo";
 export * from "./axe-scan";
 export * from "./html-validate";
 export * from "./format";
+export * from "./dates";