+++ packages/code-du-travail-frontend/src/modules/glossaire/GlossaryList.tsx
@@ -17,7 +17,7 @@ export const GlossaryList = ({ glossary = [] }: GlossaryProps) => {
     .map(({ letter }) => letter);
 
   return (
-    <ContainerWithBreadcrumbs currentPage={"Glossaire"} breadcrumbs={[]}>
+    <ContainerWithBreadcrumbs currentPage={"Glossaire"}>
       <h1 id="glossary" className={fr.cx("fr-mt-0", "fr-mb-4w")}>
         Glossaire
       </h1>